requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchmetrics>=1.0.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.1.0  # For PPO baseline comparison

# Parallelization
ray[default]>=2.9.0

# Medical Imaging
fastmri>=0.3.0
monai>=1.3.0
h5py>=3.9.0

# Image Quality Metrics
scikit-image>=0.21.0

# Experiment Tracking
wandb>=0.16.0

# Data & Utilities
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0.0

# Code Quality
flake8>=6.0.0
pytest>=7.4.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0
